Fix _rustup.tmpl
Gitpod uses (a wrapper script)[1] for rustup that doesn't seem to have a shebang when it's actually on the filesystem. This causes direct calls to "exec rustup" (not exactly that but close enough) to fail since the machine doesn't know how to run the file. This fixes that by forcing the wrapper to be run with bash, which can parse the script without any issues.

[1] https://github.com/gitpod-io/workspace-images/blob/0bc0dfd104d9a4483be235338be9c32a695e210a/chunks/lang-rust/rustup