Make sure .zshrc exits with the correct code
Assume the following code is at the end of ~/.zshrc.

  [ -f "file" ] && . "file"

If "file" does not exist, this line will exit with a non-zero exit code. Then, the prompt code will assume that exit code was the exit code of the last command. This is technically correct, but we don't actually care about this. Changing this to an `if` statement makes sure the exit code is always zero.